pyproject.toml
[project]
name = "jetCalc"
version = "0.1.0"
description = "Thermodynamic analysis for propulsion applications"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.5",
    "periodictable>=2.0.2",
    "pytest>=8.3.5",
    "pyyaml>=6.0.2",
    "ruff>=0.11.9",
]
